fixed decimal issue, still no heroku fix

+++ client/src/pages/ProductsBackEnd/ProductsBackEnd.js
@@ -42,7 +42,7 @@ class ProductsBackEnd extends Component {
             description: this.state.newProductDescription,
             category: this.state.newProductCategory,
             imgLink: this.state.newProductImgs.split(',').map(function(string){ return string.trim(); }),
-            price: parseInt(this.state.newProductPrice)
+            price: parseFloat(this.state.newProductPrice)
         };
 
         API.addProduct(newProduct)
